test: move some DateParser tests to DateRange
